update style and code BookingHistoryView
